ready project

+++ print_functions/ft_print_unsigned.c
@@ -6,13 +6,13 @@
 /*   By: jergashe <jergashe@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/01 11:01:58 by jergashe          #+#    #+#             */
-/*   Updated: 2022/11/01 17:01:30 by jergashe         ###   ########.fr       */
+/*   Updated: 2022/11/04 16:41:02 by jergashe         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../ft_printf.h"
 
-int	ft_print_unsigned(long number)
+int	ft_print_unsigned(unsigned int number)
 {
 	ft_putnbr_fd(number, 1);
 	if (number == 0)
